refactor(logger): fixed all setup_logger imports to reflect new logger.py location in utils
