+++ db.py
@@ -1,11 +1,12 @@
 import sqlite3
 from pathlib import Path
-from .logger import setup_logger
+from .utils.logger import setup_logger
 from werkzeug.security import generate_password_hash
 
 logger = setup_logger(level=0)
 BASE_DIR = Path(__file__).parent
-DATABASE = str(BASE_DIR/"db.sqlite3")
+DATABASE = str(BASE_DIR / "db.sqlite3")
+
 
 def get_database():
     logger.debug(f"Connecting to sqlite3: {DATABASE}")
@@ -15,9 +16,11 @@ def get_database():
     conn.execute("PRAGMA journal_mode=WAL;")
     return conn
 
+
 def ensure_pragmas():
     conn = get_database()
     conn.close()
+
 
 def init_db():
     logger.debug("Initializing database...")
@@ -25,7 +28,8 @@ def init_db():
     cur = conn.cursor()
 
     # USERS
-    cur.execute("""
+    cur.execute(
+        """
     CREATE TABLE IF NOT EXISTS users (
         id INTEGER PRIMARY KEY AUTOINCREMENT,
         first_name TEXT,
@@ -35,18 +39,22 @@ def init_db():
         role TEXT NOT NULL DEFAULT 'tech',
         force_password_change INTEGER NOT NULL DEFAULT 0
     );
-    """)
+    """
+    )
 
     # TECHNICIANS
-    cur.execute("""
+    cur.execute(
+        """
     CREATE TABLE IF NOT EXISTS technicians (
         id INTEGER PRIMARY KEY AUTOINCREMENT,
         name TEXT NOT NULL UNIQUE
     );
-    """)
+    """
+    )
 
     # JOBS
-    cur.execute("""
+    cur.execute(
+        """
     CREATE TABLE IF NOT EXISTS jobs (
         id INTEGER PRIMARY KEY AUTOINCREMENT,
         title TEXT NOT NULL,
@@ -75,10 +83,12 @@ def init_db():
         FOREIGN KEY (created_by) REFERENCES users(id),
         FOREIGN KEY (last_modified_by) REFERENCES users(id)
     );
-    """)
+    """
+    )
 
     # LOCKS
-    cur.execute("""
+    cur.execute(
+        """
     CREATE TABLE IF NOT EXISTS locks (
         id INTEGER PRIMARY KEY AUTOINCREMENT,
         date TEXT UNIQUE NOT NULL,
@@ -86,10 +96,12 @@ def init_db():
         locked_at TEXT DEFAULT CURRENT_TIMESTAMP,
         FOREIGN KEY (locked_by) REFERENCES users(id)
     );
-    """)
+    """
+    )
 
     # TIME OFF
-    cur.execute("""
+    cur.execute(
+        """
     CREATE TABLE IF NOT EXISTS time_off (
         id INTEGER PRIMARY KEY AUTOINCREMENT,
         technician_id INTEGER NOT NULL,
@@ -98,15 +110,18 @@ def init_db():
         reason TEXT,
         FOREIGN KEY (technician_id) REFERENCES technicians(id)
     );
-    """)
+    """
+    )
 
     # bootstrap admin if not exist
     cur.execute("SELECT COUNT(*) AS c FROM users")
     if cur.fetchone()["c"] == 0:
-        logger.warning("No users found; creating default admin (username 'admin' / password 'changeme').")
+        logger.warning(
+            "No users found; creating default admin (username 'admin' / password 'changeme')."
+        )
         cur.execute(
             "INSERT INTO users (first_name,last_name,username,password,role,force_password_change) VALUES (?,?,?,?,?,?)",
-            ("Admin", "User", "admin", generate_password_hash("changeme"), "admin", 1)
+            ("Admin", "User", "admin", generate_password_hash("changeme"), "admin", 1),
         )
 
     conn.commit()
